requirements.txt
streamlit
streamlit-lottie
plotly
pandas
requests
Pillow
numpy
scikit-learn
tensorflow
